requirements.txt
numpy>=1.18.5
pandas>=1.0.5
scikit-learn>=0.23.1
tensorflow==2.15.1
seaborn>=0.10.1
matplotlib>=3.2.2
datasets>=1.1.2
joblib>=0.15.1
flask
scikeras
gunicorn==20.1.0
requests

